apply JWT

+++ todo_app/views.py
@@ -8,27 +8,46 @@ from .serializers import TaskSerializer, TaskCreateInputSerializer, TaskUpdateIn
 
 @api_view(['GET','POST'])
 def tasks(request):
+    user = request.user
+
+    # GET Tasks
     if request.method == 'GET':
-        query_set = Task.objects.select_related('user').all()
+        if user.is_staff == True:
+            query_set = Task.objects.select_related('user').all()
+        else:
+            query_set = Task.objects.select_related('user').filter(user=user.id)
         serializer = TaskSerializer(query_set, many=True)
         return Response(serializer.data)
+
+    # ADD Task
     elif request.method == 'POST':
-        serializer = TaskCreateInputSerializer(data=request.data)
+        serializer = TaskCreateInputSerializer(data=request.data, context={'user_id': request.user.id})
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 @api_view(['GET','PUT', 'DELETE'])
 def task(request, id):
+    user = request.user
     task = get_object_or_404(Task, pk=id)
+
+    # Permission check
+    if (user.id != task.user_id):
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
+
+    # GET Task
     if request.method == 'GET':
         serializer = TaskSerializer(task)
         return Response(serializer.data)
+
+    # UPDATE Task
     elif request.method == 'PUT':
         serializer = TaskUpdateInputSerializer(task, data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data, status= status.HTTP_200_OK)
+
+    # DELETE Task
     elif request.method == 'DELETE':
         task.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
